add: loading comments
