+++ js/photo-modal.js
@@ -1,12 +1,18 @@
 import { renderComment } from './comment-list.js';
-import { onDocumentKeydown, closeBigPicture, clearCommentList, showSocialCommentCount, clearCommentsLoaderButton, onCommentsLoaderButton } from './utils/modal-util.js';
+import { isEscapeKey } from './util.js';
 
 const photoModalElement = document.querySelector('.big-picture');
 const bigPicturePhoto = photoModalElement.querySelector('.big-picture__img');
 const bigPictureSocial = photoModalElement.querySelector('.big-picture__social');
 const photoModalCloseElement = photoModalElement.querySelector('.big-picture__cancel');
 const socialCommentList = bigPictureSocial.querySelector('.social__comments');
-const commentsLoaderButton = bigPictureSocial.querySelector('.comments-loader');
+const commentLoaderButton = bigPictureSocial.querySelector('.comments-loader');
+
+const LOADING_COMMENT = 5;
+
+const clearCommentsList = () => {
+  document.querySelector('.social__comments').innerHTML = '';
+};
 
 const getPhotoData = (photo) => {
   const { url, description, likes, comments } = photo;
@@ -15,23 +21,73 @@ const getPhotoData = (photo) => {
   bigPictureSocial.querySelector('.social__caption').textContent = description;
   bigPictureSocial.querySelector('.likes-count').textContent = likes;
   bigPictureSocial.querySelector('.comments-count').textContent = comments.length;
-  clearCommentList();
+  clearCommentsList();
 };
+
+const showSocialCommentCount = (quantity) => {
+  const socialCommentCount = document.querySelector('.social__comment-count');
+  const commentsCount = document.querySelectorAll('.social__comment').length;
+  socialCommentCount.innerHTML = `${quantity} из <span class="comments-count">${commentsCount}</span> комментариев`;
+};
+
+const removeHiddenCommentClass = (element, maxValue) => {
+  for (let i = 0; i < maxValue; i++) {
+    element[i].classList.remove('hidden');
+  }
+};
+
+const onCommentsLoaderButton = () => {
+  const hiddenComments = document.querySelectorAll('.social__comment.hidden');
+  const commentsCount = document.querySelectorAll('.social__comment').length;
+  const hiddenCommentsLength = hiddenComments.length;
+  if (hiddenCommentsLength <= LOADING_COMMENT) {
+    commentLoaderButton.classList.add('hidden');
+  }
+  if (hiddenCommentsLength < LOADING_COMMENT) {
+    removeHiddenCommentClass(hiddenComments, hiddenCommentsLength);
+  } else {
+    removeHiddenCommentClass(hiddenComments, LOADING_COMMENT);
+  }
+  const restHiddenComments = document.querySelectorAll('.social__comment.hidden').length;
+  showSocialCommentCount(commentsCount - restHiddenComments);
+};
+
+const clearCommentsLoaderButton = () => {
+  commentLoaderButton.removeEventListener('click', onCommentsLoaderButton);
+  commentLoaderButton.classList.remove('hidden');
+};
+
+const onDocumentKeydown = (evt) => {
+  if (isEscapeKey(evt)) {
+    evt.preventDefault();
+    closeBigPicture();
+    clearCommentsLoaderButton();
+  }
+};
+
+function closeBigPicture() {
+  document.querySelector('.big-picture').classList.add('hidden');
+  document.body.classList.remove('modal-open');
+  document.removeEventListener('keydown', onDocumentKeydown);
+  clearCommentsList();
+}
 
 const showComment = (photo) => {
   photo.comments.forEach((comment, index) => {
     socialCommentList.append(renderComment(comment));
-    if (index > 4) {
+    if (index > LOADING_COMMENT - 1) {
       socialCommentList.querySelectorAll('li')[index].classList.add('hidden');
     }
   });
-  if (photo.comments.length < 6) {
-    commentsLoaderButton.classList.add('hidden');
+  if (photo.comments.length <= LOADING_COMMENT) {
+    commentLoaderButton.classList.add('hidden');
     showSocialCommentCount(photo.comments.length);
+  } else {
+    showSocialCommentCount(LOADING_COMMENT);
   }
 };
 
-const loadComment = () => commentsLoaderButton.addEventListener('click', (onCommentsLoaderButton));
+const loadComment = () => commentLoaderButton.addEventListener('click', onCommentsLoaderButton);
 
 const openModalBigPicture = (thumbnail, photo) => {
   thumbnail.addEventListener('click', (evt) => {
@@ -48,7 +104,7 @@ const openModalBigPicture = (thumbnail, photo) => {
 const closeModalBigPicture = () => {
   photoModalCloseElement.addEventListener('click', () => {
     closeBigPicture();
-    clearCommentsLoaderButton(commentsLoaderButton);
+    clearCommentsLoaderButton();
   });
 };
